Move core tags-related code inside features/tags directory
