+++ src/features/settings/components/SettingsFormSidebar.svelte
@@ -9,7 +9,7 @@ import {
   SETTINGS_SHORTCUTS,
   SETTINGS_MISCELLANEOUS,
 } from '@lib/constants';
-import { tags } from '@stores/tags';
+import { tags } from '@features/tags/store';
 
 export let value;
