+++ src/features/todos/components/TodoForm.svelte
@@ -2,14 +2,14 @@
 import { createEventDispatcher, onMount, onDestroy } from 'svelte';
 import orderBy from 'lodash/orderBy';
 
-import { tags } from '@stores/tags';
 import { sanitizeText, unsanitizeText } from '@lib/sanitize';
 import { enableShortcut, disableShortcut } from '@lib/shortcuts';
 import { TODOS_TODAY, TODOS_THIS_WEEK, TODOS_EVENTUALLY } from '@lib/constants';
+import { tags } from '@features/tags/store';
 
 import Selector from '@components/Selector.svelte';
 import Button from '@components/Button.svelte';
-import TagsInput from '@components/TagsInput.svelte';
+import TagsInput from '@features/tags/components/TagsInput.svelte';
 
 export let data = {
   list: TODOS_EVENTUALLY,
